backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
pymongo>=4.6.0
sqlalchemy>=2.0.0
boto3>=1.28.0
requests>=2.31.0
gunicorn>=21.2.0
pytest>=7.4.0
PyJWT>=2.8.0
bcrypt>=4.1.0
pydantic>=2.0.0
